pyproject.toml
# Project configuration for ForTheRekord - Rekordbox to Spotify synchronization tool

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fortherekord"
version = "0.1.0"
description = "Synchronize Rekordbox DJ library with Spotify playlists"
authors = [{name = "ForTheRekord", email = "user@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "pydantic>=2.0.0", 
    "pyyaml>=6.0.0",
    "spotipy>=2.22.0",
    "rapidfuzz>=3.0.0",
    "requests>=2.28.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[project.scripts]
fortherekord = "fortherekord.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
